wired

+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/movement/LongJump.kt
@@ -114,7 +114,7 @@ class LongJump : Module() {
                 airTicks = 0
             }
             if(mc.thePlayer.isInWater) {
-                mc.thePlayer.motionY = jartexYValue.get()
+                mc.thePlayer.motionY = jartexYValue.get().toDouble()
                 MovementUtils.strafe(jartexHValue.get())
             }
         }
